Add altair processBlock perf tests
