+++ packages/beacon-state-transition/test/perf/phase0/block/processBlock.test.ts
@@ -1,0 +1,123 @@
+import {itBench, setBenchOpts} from "@dapplion/benchmark";
+import {
+  ACTIVE_PRESET,
+  MAX_ATTESTATIONS,
+  MAX_ATTESTER_SLASHINGS,
+  MAX_DEPOSITS,
+  MAX_PROPOSER_SLASHINGS,
+  MAX_VOLUNTARY_EXITS,
+  PresetName,
+} from "@chainsafe/lodestar-params";
+import {allForks} from "../../../../src";
+import {generatePerfTestCachedStatePhase0, perfStateId} from "../../util";
+import {getBlockPhase0} from "./util";
+
+// As of Jun 12 2021
+// Process block
+// ================================================================
+// Process block with 0 validator exit                                    233.6434 ops/s      4.280027 ms/op   3491 runs    15.01 s
+// Process block with 1 validator exit                                    41.33581 ops/s      24.19210 ms/op    619 runs    15.00 s
+// Process block with 16 validator exits                                  42.34492 ops/s      23.61558 ms/op    635 runs    15.02 s
+
+// Processing a block consist of three steps
+// 1. Verifying signatures
+// 2. Running block processing (state transition function)
+// 3. Hashing the state
+//
+// Performance cost of each block depends on the size of the state + the operation counts in the block
+//
+//
+// ### Verifying signatures
+// Signature verification is done in bulk using batch BLS verification. Performance is proportional to the amount of
+// sigs to verify and the cost to construct the signature sets from TreeBacked data.
+//
+// - Proposer sig:           1 single
+// - RandaoReveal sig:       1 single
+// - ProposerSlashings sigs: ops x 2 single
+// - AttesterSlashings sigs: ops x 2 agg (90 bits on avg)
+// - Attestations sigs:      ops x 1 agg (90 bits on avg)
+// - VoluntaryExits sigs:    ops x 1 single
+// - Deposits sigs:          ops x 1 single
+//
+// A mainnet average block has:
+//   1 + 1 + 90 = 92 sigs
+//   90 * 90 = 8100 pubkey aggregations
+// A maxed block would have:
+//   1 + 1 + 16 * 2 + 2 * 2 + 128 + 16 + 16 = 198 sigs
+//   2 * 2 * 128 + 128 * 128 = 16896 pubkey aggregations
+//
+//
+// ### Running block processing
+// Block processing is relatively fast, most of the cost is reading and writing tree data. The performance of
+// processBlock is properly tracked with this performance test.
+//
+// - processBlockHeader        : -
+// - processRandao             : -
+// - processEth1Data           : -
+// - processOperations         --
+//   - processProposerSlashing : -
+//   - processAttesterSlashing : -
+//   - processAttestation      : -
+//   - processDeposit          : -
+//   - processVoluntaryExit    : -
+//
+// ### Hashing the state
+// Hashing cost is dependant on how many nodes have been modified in the tree. After mutating the state, just count
+// how many nodes have no cached _root, then multiply by the cost of hashing.
+//
+
+describe("Process block", () => {
+  setBenchOpts({
+    maxMs: 60 * 1000,
+    minMs: 15 * 1000,
+    runs: 1024,
+  });
+
+  if (ACTIVE_PRESET !== PresetName.mainnet) {
+    throw Error(`ACTIVE_PRESET ${ACTIVE_PRESET} must be mainnet`);
+  }
+
+  const baseState = generatePerfTestCachedStatePhase0() as allForks.CachedBeaconState<allForks.BeaconState>;
+
+  const worstCaseBlockState = baseState.clone();
+  const worstCaseBlock = getBlockPhase0(worstCaseBlockState, {
+    proposerSlashingLen: MAX_PROPOSER_SLASHINGS,
+    attesterSlashingLen: MAX_ATTESTER_SLASHINGS,
+    attestationLen: MAX_ATTESTATIONS,
+    depositsLen: MAX_DEPOSITS,
+    voluntaryExitLen: MAX_VOLUNTARY_EXITS,
+    bitsLen: 128,
+  });
+
+  const averageMainnetBlockState = baseState.clone();
+  const averageMainnetBlock = getBlockPhase0(averageMainnetBlockState, {
+    proposerSlashingLen: 0,
+    attesterSlashingLen: 0,
+    attestationLen: 90,
+    depositsLen: 0,
+    voluntaryExitLen: 0,
+    bitsLen: 90,
+  });
+
+  itBench(
+    {id: `processBlock phase0 ${perfStateId} - maxed ops`, beforeEach: () => worstCaseBlockState.clone()},
+    (state) => {
+      allForks.stateTransition(state as allForks.CachedBeaconState<allForks.BeaconState>, worstCaseBlock, {
+        verifyProposer: false,
+        verifySignatures: false,
+        verifyStateRoot: false,
+      });
+    }
+  );
+
+  itBench(
+    {id: `processBlock phase0 ${perfStateId} - average`, beforeEach: () => averageMainnetBlockState.clone()},
+    (state) => {
+      allForks.stateTransition(state as allForks.CachedBeaconState<allForks.BeaconState>, averageMainnetBlock, {
+        verifyProposer: false,
+        verifySignatures: false,
+        verifyStateRoot: false,
+      });
+    }
+  );
+});
